Update Binary Length Checker
Remove the comment about not using strlen() directly due to null-bytes, and add tests to specifically check IPs that contain null-bytes at the start and end of the binary sequence. This doesn't seem to be needed anymore (mostly a bug in a previous version of PHP that is no longer supported).